+++ src/components/common/atoms/CarouselSlide.jsx
