+++ src/components/layouts/mainLayout/organisms/BottomNavBar.jsx
@@ -6,7 +6,7 @@ import {HiOutlineTicket} from "react-icons/hi";
 
 const BottomNavBar = ({ activatedTab }) => {
   return (
-    <nav className="bottom-nav-bar fixed bottom-0 flex w-full min-w-[400px] max-w-[800px] justify-around bg-white py-2 shadow-2xl rounded-t-2xl drop-shadow divide-x divide-gray-300 ">
+    <nav className="bottom-nav-bar width-flex-layout fixed bottom-0 flex justify-around bg-white py-2 shadow-2xl rounded-t-2xl drop-shadow divide-x divide-gray-300 ">
       <NavItem id={1} to={"/"} icon={<GrMap size={30} />} label={"Spots"} />
       <NavItem id={2} to={"/foods"} icon={<BiFoodMenu size={30} />} label={"Foods"} />
       <NavItem id={3} to={"/userinfo/wishlist/restaurant"} icon={<AiOutlineHeart size={30} />} label={"Wishlist"} />
