+++ src/components/pages/homePage/atoms/ListTitle.jsx
@@ -1,0 +1,7 @@
+const ListTitle = ({ children }) => {
+  return (
+      <h2 className={"list-title px-2 py-1 text-3xl font-extrabold"}>{children}</h2>
+  );
+};
+
+export default ListTitle;
