+++ src/components/pages/homePage/atoms/PlaceCard.jsx
