+++ src/components/pages/homePage/molecules/FestivalCardList.jsx
@@ -1,13 +1,13 @@
 import FestivalCard from "../atoms/FestivalCard";
-import HorizontalListWrapper from "../atoms/HorizontalListWrapper";
+import HorizontalListWrapper from "../../../common/atoms/HorizontalListWrapper";
 
 const FestivalCardList = ({ festivals }) => {
   return (
     <HorizontalListWrapper>
-      {festivals.map((festival) => {
+      {festivals.map((festival, index) => {
         return (
           <FestivalCard
-            key={festival.id}
+            key={index}
             image={festival.image}
             alt={`poster of ${festival.name}`}
             to={`festival/${festival.id}`}
