+++ src/components/pages/homePage/molecules/RestaurantCardList.jsx
@@ -1,13 +1,13 @@
-import HorizontalListWrapperMediaQuery from "../atoms/HorizontalListWrapperMediaQuery";
+import HorizontalListWrapperMediaQuery from "../../../common/atoms/HorizontalListWrapperMediaQuery";
 import PlaceCard from "../atoms/PlaceCard";
 
 const RestaurantCardList = ({ restaurants }) => {
   return (
     <HorizontalListWrapperMediaQuery>
-      {restaurants.map((restaurant) => {
+      {restaurants.map((restaurant, index) => {
         return (
           <PlaceCard
-            key={restaurant.id}
+            key={index}
             image={restaurant.image}
             alt={`poster of ${restaurant.name}`}
             to={`restaurant/${restaurant.id}`}
