+++ src/components/pages/searchPage/assets/map-icon.png
