--- src/pages/homePage/atoms/ListTitle.jsx
@@ -1,7 +1,0 @@
-const ListTitle = ({ children }) => {
-  return (
-      <h2 className={"list-title px-2 text-3xl font-extrabold"}>{children}</h2>
-  );
-};
-
-export default ListTitle;
